+++ rust/otel-arrow-rust/src/validation/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod collector;
+mod otap;
 mod otlp;
 mod scenarios;
 mod service_type;
